+++ server/controllers/user_controller.js
@@ -1,6 +1,7 @@
 const User = require('../models/users');
 const Ticket = require('../models/tickets');
 const bcrypt = require('bcrypt');
+const { check, validationResult } = require('express-validator')
 
 /**
  * Gets all users from the DB.
@@ -40,7 +41,7 @@ let user_create = function (req, res) {
     let user = new User(
         {
             username: body.username,
-            password: bcrypt.hashSync(body.password, 10),
+            password: body.password, //bcrypt.hashSync(body.password, 10)
             firstName: body.firstName,
             lastName: body.lastName,
             dni: body.dni,
@@ -172,6 +173,58 @@ let user_tickets = function (req, res) {
     })
 }
 
+/**
+ * Autenticates User.
+ * @module user
+ * @function
+ * @param {Object} req - Express request object
+ * @param {Object} res - Express response object
+ * @return {Object} res .
+ */
+let login = function (req, res) {
+    var errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+        req.session.errors = errors.array();
+        req.session.success = false;
+    } else {
+        //Esto no esta funcionando, no se como pasar el id. 
+        User.findOne({username: req.body.username}, (err, user) => { // chequear User = null
+            if (err) {
+                req.session.success = false;
+                return res.status(400).json({
+                    ok: false,
+                    err
+                })
+            }
+
+            if (!user) {
+                req.session.success = false;
+                return res.status(400).json({
+                    ok: false,
+                    err: {
+                        message: 'Usuario o contraseña incorrectos' // Usuario en este caso
+                    }
+                });
+            }
+
+            if (user.password === req.body.password){ //bcrypt.hashSync(req.params.password, 10)
+                req.session.success = true;
+                return res.status(200).json({
+                    ok: true,
+                    user: user.toJSON(),
+                })
+            } else {
+                req.session.success = false;
+                return res.status(401).json({
+                    ok: false,
+                    error: 'Usuario o contraseña incorrectos'
+                })
+            }  
+        })
+    }
+}
+
 
 module.exports = {
     user_all: user_all,
@@ -180,4 +233,5 @@ module.exports = {
     user_delete_by_id: user_delete_by_id,
     user_update_by_id: user_update_by_id,
     user_tickets: user_tickets,
+    login: login
 }
